feat (MircofacetMaterial) :

+++ src/Core/FunctionLayer/Sampler.h
@@ -59,7 +59,6 @@ inline  Vec3d SquareToUniformSphere(const Point2d &sample) {
 }
 
 inline  float SquareToUniformSpherePdf(const Vec3d &v) {
-//    throw NoriException("SquareToUniformSpherePdf() is not yet implemented!");
     return 0.25f*INV_PI;
 }
 
@@ -68,12 +67,10 @@ inline  Vec3d SquareToUniformHemisphere(const Point2d &sample) {
     float r = std::sqrt(std::max((float )0, (float)1 - z * z));
     float phi = 2 * M_PI * sample[1];
     return {r * std::cos(phi), r * std::sin(phi), abs(z)};
-    //throw NoriException("SquareToUniformHemisphere() is not yet implemented!");
 }
 
 inline float SquareToUniformHemispherePdf(const Vec3d &v) {
     return v[2] >=0 ? 0.5f * INV_PI : .0f;
-//    throw NoriException("SquareToUniformHemispherePdf() is not yet implemented!");
 }
 
 inline  Vec3d SquareToCosineHemisphere(const Point2d &sample) {
@@ -81,24 +78,22 @@ inline  Vec3d SquareToCosineHemisphere(const Point2d &sample) {
     float phi=sample.y*2*M_PI;
 
     return {sqrt(sample.x)* cos(phi),sqrt(sample.x)*sin(phi),z};
-//    throw NoriException("SquareToCosineHemisphere() is not yet implemented!");
 }
 
 inline  float SquareToCosineHemispherePdf(const Vec3d &v) {
     return v[2] >=0 ? v.z * INV_PI : .0f;
 }
 //
-inline  Vec3d SquareToBeckmann(const Point2d &sample, float alpha) {
+inline  Vec3d SquareToBeckmann(const Point2d &sample,double alpha) {
     auto tan2theta= -alpha*alpha*log( sample.x );
     auto cosTheta=sqrt(1/(1+tan2theta));
     auto sinTheta= sqrt(1-cosTheta*cosTheta);
     auto phi=sample.y * 2 * M_PI;
     Vec3d t1= Vec3d(sinTheta*cos(phi), sinTheta*sin(phi),cosTheta);
     return t1;
-//    throw NoriException("SquareToBeckmann() is not yet implemented!");
 }
 //
-inline  float SquareToBeckmannPdf(const Vec3d &m, float alpha) {
+inline  float SquareToBeckmannPdf(const Vec3d &m, double alpha) {
     if(m.z<=0)
         return 0.0f;
     auto cosTheta=m.z;
